Fixed compilation on 64-bit Linux
An int is generally 32-bit, while a long could be 64-bit, so the typedef
for u32 and s32 has been changed to use int. This fixes the assertion on
startup.

A lot of places in the code assumed that it could cast a pointer to a
32-bit integer and back. For this purpose, two new integer typedefs have
been introduced, uintptr and sintptr, which are now used in all these
cases. They are typedefs for unsigned long and long respectively, for
lack of something better for now.

Some printf calls were casting a pointer to a 32-bit integer and then
formatting it like a pointer value. These calls have been changed to use
%p, which prints a pointer value in this format.